Add parens to zip output used as for loop args
The zip operator now returns an unflattened Parcel of pairs, which is not
what was expected as input to the for loop's body.  The solution is simply
to add parentheses around the arguments to the body of the for loop in order
to extract the correct data pairs of elements.  This change makes the test
suite pass again.